# docs/swagger.yaml
definitions:
...
  dto.CreateMenu:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
    required:
    - name
    - path
    type: object
...
  dto.DeleteMenuBulk:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  dto.MenuAPI:
    properties:
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      type:
        type: string
    type: object
...
  dto.PayloadCreateMenu:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CreateMenu'
        minItems: 1
        type: array
    required:
    - data
    type: object
...
  dto.UpdateMenu:
    properties:
      apis:
        items:
          type: integer
        type: array
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
    required:
    - name
    - path
    type: object
...
  middlewares.ResponseError:
    properties:
      message: {}
      request_id:
        type: string
      time:
        type: string
    type: object
  models.Action:
    properties:
      api:
        type: string
      created_at:
        type: string
      default:
        type: boolean
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ActionsMenus:
    properties:
      actionId:
        type: integer
      actions:
        $ref: '#/definitions/models.Action'
      created_at:
        type: string
      id:
        type: integer
      menuId:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Menu:
    properties:
      actions_menus:
        items:
          $ref: '#/definitions/models.ActionsMenus'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
...
info:
  contact:
    name: PT Tata Kelola Dokumen
  description: Aplikasi Assessment & Monitoring
  license:
    name: Private License
  title: EJM Assessment & Monitoring API Documentation
  version: "1.0"
paths:
...
  /menus:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuAPI'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Ambil semua data menu
      tags:
      - Menu
  /menus/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID Menu
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Delete Menu
      tags:
      - Menu
    get:
      consumes:
      - application/json
      parameters:
      - description: ID Menu
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Detail Menu
      tags:
      - Menu
    put:
      consumes:
      - application/json
      parameters:
      - description: ID Menu
        in: path
        name: id
        required: true
        type: integer
      - description: body update
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Update Menu
      tags:
      - Menu
  /menus/bulk:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body delete
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteMenuBulk'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Delete Menu Bulk
      tags:
      - Menu
  /menus/create-bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Data Menu
        in: body
        name: menus
        required: true
        schema:
          $ref: '#/definitions/dto.PayloadCreateMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Create Menu Bulk
      tags:
      - Menu
  /menus/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Halaman
        in: query
        name: page
        required: true
        type: string
      - description: Jumlah Data Per halaman
        in: query
        name: page_size
        required: true
        type: string
      - description: Mencari Data
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Ambil semua data menu
      tags:
      - Menu
...
schemes:
- http
swagger: "2.0"